+++ clang-tools-extra/test/clang-tidy/checkers/performance/constexpr-non-static-in-scope.cpp
@@ -1,0 +1,14 @@
+// RUN: %check_clang_tidy %s performance-constexpr-non-static-in-scope %t
+
+// FIXME: Add something that triggers the check here.
+void f();
+// CHECK-MESSAGES: :[[@LINE-1]]:6: warning: function 'f' is insufficiently awesome [performance-constexpr-non-static-in-scope]
+
+// FIXME: Verify the applied fix.
+//   * Make the CHECK patterns specific enough and try to make verified lines
+//     unique to avoid incorrect matches.
+//   * Use {{}} for regular expressions.
+// CHECK-FIXES: {{^}}void awesome_f();{{$}}
+
+// FIXME: Add something that doesn't trigger the check here.
+void awesome_f2();